backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ibm-watson-machine-learning==1.0.312
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
numpy==1.24.3
pandas==1.5.3
ibm-watsonx-ai
pymupdf

